requirements.txt
bleak>=0.22
opencv-python-headless>=4.9
numpy>=1.26
openai>=1.40
pyserial>=3.5
svgpathtools>=1.6
flask>=3.0
python-dotenv>=1.0
